+++ src/ontodia/data/model.ts
@@ -36,57 +36,3 @@ export interface LinkCount {
 export interface LinkType extends LinkCount {
     label: { values: LocalizedString[] };
 }
-
-export interface LinkLabelStyle {
-    position?: number;
-    attrs?: {
-        rect?: {
-            fill?: string;
-            'stroke'?: string;
-            'stroke-width'?: number;
-        };
-        text?: {
-            fill?: string;
-            'stroke'?: string;
-            'stroke-width'?: number;
-        };
-    };
-}
-
-export interface LinkStyle {
-    attrs?: {
-        '.connection'?: {
-            fill?: string;
-            stroke?: string;
-            'stroke-width'?: number;
-        },
-        '.marker-source'?: {
-            fill?: string;
-            stroke?: string;
-            'stroke-width'?: number;
-            d?: string;
-        },
-        '.marker-target'?: {
-            fill?: string;
-            stroke?: string;
-            'stroke-width'?: number;
-            d?: string;
-        }
-    };
-    labels?: LinkLabelStyle[];
-    connector?: {
-        name?: string;
-        args?: {
-            radius?: number;
-        };
-    };
-    router?: {
-        name?: string;
-        args?: {
-            startDirections?: string[];
-            endDirections?: string[];
-            excludeTypes?: string[];
-        };
-    };
-    z?: number;
-}
